% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc}
\alias{abc}
\title{Calculate a, b, and c}
\usage{
abc(var_epsilon, var_v)
}
\arguments{
\item{var_epsilon}{Variance of epsilon}

\item{var_v}{Variance of V}
}
\value{
A list containing a, b, and c
}
\description{
Calculates the constants of the posterior distribution of gf.
}
