% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapr_stage_2.R
\name{calculate_beta}
\alias{calculate_beta}
\title{Calculate beta coefficients based on model type}
\usage{
calculate_beta(model_type, coefficients, posterior)
}
\arguments{
\item{model_type}{The type of model ("lm", "probit" or "cox")}

\item{coefficients}{A list containing gamma and theta coefficients}

\item{posterior}{The posterior values from the abc function}
}
\value{
A named numeric vector of beta coefficients
}
\description{
Calculate beta coefficients based on model type
}
