% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapr_wrapper.R
\name{hapr}
\alias{hapr}
\title{Heritability adjusted prediction}
\usage{
hapr(y, gc, w, model_type, improvement_ratio)
}
\arguments{
\item{y}{Outcome variable}

\item{gc}{Polygenic risk score (has to be normalized)}

\item{w}{Control variables}

\item{model_type}{Type of model to fit ("lm" or "probit")}

\item{improvement_ratio}{The ratio of R-squared of the future fit to the current fit.}
}
\value{
A hapr_fit object.
}
\description{
Fits the HARP model given the outcome y, PRS gc, and control variables w.
The improvement ratio specifies the expected increase in R-squared of the
future GWAS relative to the current GWAS.
}
