% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{hapr_cox}
\alias{hapr_cox}
\title{Heritability adjusted prediction for the Cox proportional hazards model}
\usage{
hapr_cox(
  y,
  gc,
  w,
  improvement_ratio = NULL,
  r2_current = NULL,
  r2_future = NULL
)
}
\arguments{
\item{y}{Outcome variable: a Surv object compatible with survival::coxph()}

\item{gc}{Polygenic risk score (has to be normalized)}

\item{w}{Control variables}

\item{improvement_ratio}{The ratio of R-squared of the future fit to the current fit.}

\item{r2_current}{The R-squared of the current fit}

\item{r2_future}{The R-squared of the future fit}
}
\value{
A hapr_cox_fit object.
}
\description{
Fits the HARP model given the outcome y, PRS gc, and control variables w.
}
