% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_second_stage.R
\name{hapr_cox_second_stage}
\alias{hapr_cox_second_stage}
\title{HAPR Cox proportional hazards model second stage fit}
\usage{
hapr_cox_second_stage(
  first_stage,
  improvement_ratio = NULL,
  r2_current = NULL,
  r2_future = NULL
)
}
\arguments{
\item{first_stage}{A hapr_cox_first_stage_fit object}

\item{improvement_ratio}{The ratio to extrapolate by}

\item{r2_current}{The R-squared of the current fit}

\item{r2_future}{The R-squared of the future fit}
}
\value{
A hapr_cox_fit object containing the results of the second stage
}
\description{
After fitting the first stage, we can specify an improvement ratio to estimate
the full model. Alternatively, we can specify the R-squared of the future fit
to estimate the improvement ratio. If specifying the r2_future, you can also
specify the r2_current to estimate the improvement ratio. Otherwise, the
r2_current is extracted from the data in the first stage fit.
}
