% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapr_stage_1.R
\name{hapr_first_stage}
\alias{hapr_first_stage}
\title{HAPR first stage fit}
\usage{
hapr_first_stage(y, gc, w, model_type)
}
\arguments{
\item{y}{Outcome variable}

\item{gc}{Polygenic risk score (has to be normalized)}

\item{w}{Control variables}

\item{model_type}{"lm", "probit", or "cox"}
}
\value{
A hapr_first_stage_fit object containing the results of the first stage.
}
\description{
Fits the first stage of the HARP model given the outcome y, PRS gc, and control variables w.
}
\details{
This returns a first stage fit, which does not need to assume an improvement ratio. Run
hapr_second_stage(first_stage_fit, improvement_ratio) to specify an improvement ratio and get the full model.
}
