% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{hapr_probit}
\alias{hapr_probit}
\title{Heritability adjusted prediction for probit models}
\usage{
hapr_probit(
  y,
  gc,
  w,
  improvement_ratio = NULL,
  r2_current = NULL,
  r2_future = NULL
)
}
\arguments{
\item{y}{Outcome variable}

\item{gc}{Polygenic risk score (has to be normalized)}

\item{w}{Control variables}

\item{improvement_ratio}{The ratio of R-squared of the future fit to the current fit.}

\item{r2_current}{The R-squared of the current fit}

\item{r2_future}{The R-squared of the future fit}
}
\value{
A hapr_probit_fit object.
}
\description{
Fits the HARP model given the outcome y, PRS gc, and control variables w.
This is a convenience function that combines the first and second stage fits.
}
