% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_first_stage.R
\name{hapr_probit_first_stage}
\alias{hapr_probit_first_stage}
\title{HAPR probit model first stage fit}
\usage{
hapr_probit_first_stage(y, gc, w)
}
\arguments{
\item{y}{Outcome variable}

\item{gc}{Polygenic risk score (has to be normalized)}

\item{w}{Control variables}
}
\value{
A hapr_probit_first_stage_fit object containing the results of the first stage.
}
\description{
HAPR probit model first stage fit
}
\details{
Fits the HARP model given the outcome y, PRS gc, and control variables w. This returns
a first stage fit, which does not need to assume an improvement ratio. Run
hapr_probit_second_stage(first_stage_fit, improvement_ratio) to specify an improvement ratio and get the full model.
}
