% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{hapr_simulate}
\alias{hapr_simulate}
\title{Simulate data from a hapr_fit model}
\usage{
hapr_simulate(object, w, gc = NULL, repetitions = 1)
}
\arguments{
\item{object}{A hapr_fit or hapr_mle_fit object}

\item{w}{A data frame or numeric matrix of control variables/covariates}

\item{gc}{Optional vector of polygenic risk scores. If NULL, they will be simulated
based on the parameters in the model}

\item{repetitions}{Number of times to repeat each row in w}
}
\value{
A data frame containing:
\itemize{
\item gf: Simulated true genetic value
\item gc: Simulated or provided polygenic risk score
\item Additional columns from the w data frame
}
}
\description{
This function simulates data from a fitted HAPR model. It generates genetic scores
and outcomes based on the model parameters estimated in the hapr_fit object.
}
\details{
The function uses the parameters from the HAPR model to simulate genetic data and outcomes.
When gc is not provided, both the true genetic effect (gf) and measured genetic score (gc)
are simulated according to the heritability parameters in the model. If gc is provided,
the true genetic effect (gf) is simulated.
}
\examples{
\dontrun{
# Fit a HAPR model
fit <- hapr(y = outcome, gc = genetic_score, w = covariates, 
            model_type = "lm", improvement_ratio = 2)
            
# Simulate 100 observations using the same covariates
sim_data <- hapr_simulate(fit, w = covariates, repetitions = 10)

# Simulate with provided genetic scores
sim_data2 <- hapr_simulate(fit, w = covariates, gc = genetic_score)
}

}
\seealso{
\code{\link{hapr}}, \code{\link{hapr_first_stage}}, \code{\link{hapr_second_stage}}
}
