% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-methods.R
\name{hapr_survfit}
\alias{hapr_survfit}
\title{Create a survival fit from a HAPR Cox model}
\usage{
hapr_survfit(fit, covariates = "gf_w", newdata)
}
\arguments{
\item{fit}{A hapr_fit object from a Cox model}

\item{covariates}{Character string indicating which covariates to use, one of: 'gf_w' (default), 'gc_w', or 'w'}

\item{newdata}{A data frame containing new observations for which to compute survival curves}
}
\value{
A hapr_survfit object containing survival probabilities over time
}
\description{
Create a survival fit from a HAPR Cox model
}
