% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapr_methods.R
\name{print.summary.hapr_fit}
\alias{print.summary.hapr_fit}
\title{Print method for summary.hapr_fit objects}
\usage{
\method{print}{summary.hapr_fit}(x, ...)
}
\arguments{
\item{x}{A summary.hapr_fit object}

\item{...}{Additional arguments (not used)}
}
\value{
The input object, invisibly
}
\description{
Print method for summary.hapr_fit objects
}
