% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strippers.R
\name{strip_lm}
\alias{strip_lm}
\title{Strip a linear model to essential components}
\usage{
strip_lm(fit)
}
\arguments{
\item{fit}{An object of class 'lm' (linear model)}
}
\value{
A list containing:
\itemize{
\item coefficients: Model coefficients
\item vcov_coefficients: Variance-covariance matrix of coefficients
\item r2: R-squared value of the model
\item explained_variance: Variance of fitted values
\item sigma_squared: Residual variance
\item var_outcome: Variance of the outcome variable
\item stripped_model: A trimmed version of the original model object
}
}
\description{
This internal function takes a linear model object and extracts the essential
components while removing memory-intensive parts of the model.
}
\keyword{internal}
