% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strippers.R
\name{strip_probit}
\alias{strip_probit}
\title{Strip a probit model to essential components}
\usage{
strip_probit(fit)
}
\arguments{
\item{fit}{An object of class 'glm' with probit link function}
}
\value{
A list containing:
\itemize{
\item coefficients: Model coefficients
\item vcov_coefficients: Variance-covariance matrix of coefficients
\item r2: R-squared value on the liability scale
\item stripped_model: A trimmed version of the original model object
}
}
\description{
This internal function takes a probit GLM object and extracts the essential
components while removing memory-intensive parts of the model.
}
\keyword{internal}
