% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapr_stage_2.R
\name{hapr_second_stage}
\alias{hapr_second_stage}
\title{HAPR second stage fit}
\usage{
hapr_second_stage(first_stage, improvement_ratio)
}
\arguments{
\item{first_stage}{A hapr_first_stage_fit object}

\item{improvement_ratio}{The ratio to extrapolate by (required)}
}
\value{
A hapr_lm_fit object containing the results of the second stage
}
\description{
Fits the full HARP model given the first stage fit and an improvement ratio.
Uses manual Delta Method for standard errors.
}
