% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapr_methods.R
\name{print.hapr_fit}
\alias{print.hapr_fit}
\title{Print method for hapr_fit objects}
\usage{
\method{print}{hapr_fit}(x, ...)
}
\arguments{
\item{x}{A hapr_fit object}

\item{...}{Additional arguments (not used)}
}
\value{
The input object, invisibly
}
\description{
Prints a concise summary of a hapr_fit object.
}
